refactor: DTO 액터별로 분리
